Added Binning

+++ app/src/main/java/physicsengine/Common.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class Common {
 
-    public static int constrain(int val, int min_val, int max_val) {
+    public static double constrain(int val, int min_val, int max_val) {
         return Math.min(max_val, Math.max(min_val, val));
     }
